Introduce the ConfigPolicy CRD.
This introduces the ConfigPolicy CRD. This will be used by the
Microservice to set up correct configuration objects for the associated
containers.

The controller will monitor the values of `Env` and `EnvFrom` and store
them sorted and hashed in the `Hashed` status value. If this value
changes, the `LastUpdated` value will be refreshed. This is an
indication for the Microservice controller to deploy a new version of
the Microservice.