Introduce EnvFromSource hash functionality.
This will now load entire secrets and config maps and sort the keys
accordingly to then hash the values.

This also fixes an issue where we prepend the key to the hashed value.
This makes sure that when someone updates a key reference, the secret
gets properly updated as well.