Microservice Controller deploys multiple versions.
The Microservice Controller should be taking care of deploying multiple
versions of a VersionedMicroservice. These versions need to be defined
by the ImagePolicy which we then pull in.

By storing the released versions on the Microservice CRD, we also have
an overview of what is actually being deployed to the cluster and what
not. Other controllers can use this information to create things like
Services and Ingresses which are based on the labels we use for these
releases.