Move NetworkPolicy to separate controller.
The NetworkPolicy will get it's own functionality. Therefore, we need to
move it into it's own controller.

It will live on it's own, controlling the services and ingresses for a
specific microservice. It will know about how to set up internal and
external DNS and how to roll over to newer versions.