Use ShortHash for naming schema.
When concatenating the version and build, it could be that we use
characters that are not allowed for DNS names. To avoid this, we need to
hash this value.

This uses a 10 character hash value prefixed by the CRD name. This also
fixes an issue where we previously used the NetworkPolicy name as the
Service DNS prefix. We will now use the Microservice name.