Link OwnerReference to release and svc/ing.
By linking the OwnerReference to a specific release, we know which
VersionedMicroservice owns the release. This way we can use this to
propagate it across other elements in our cluster like Services and
Ingresses.

By adding the OwnerReference correctly, we don't have to implement code
to delete these Services and Ingresses.