Correct Syncing behaviour.
Kubekit has a patch which fixes an issue where it would try and update
an object with it's own status, causing an infinite update loop.

By updating kubekit, we can now drop the ShouldSync check. This will
give us a better experience by:

1) no need for the ShouldSync checks, which are fragile,
2) the underlying parts of a sync action are now triggered correctly as
well, this means that the cluster will always be in a state which is
correct. Before, this could have been different