ObjectReference instead of LocalObjectReference.
Some policies can be shared across multiple environments. We should
allow users to specify this as such.

By using `ObjectReference` instead of `LocalObjectReference`, a user can
define the namespace an object lives in. This is useful for fetching
data across multiple namespaces.

The caveat here is that with RBAC enabled, users must set up access
properly.