github repository status syncs across multiple networkpolices
Because of this, a single networkpolicy can't be treated as the lone
source of truth for ghr release status. Change the ghr controller to
only update relevant releases when a networkpolicy changes.

We will still need to perform a full reconciliation on the github
repository to correct any times we've failed to reconcile on a network
policy change.