Connect label matching to image policy filtering
The image policy match section can now optionally support a labels
field, which, if specified, takes a series of image labels.

If any are supplied, all labels must exist on an image, and the value
must match the mapping, for an image to match. Both a name match and
label matches can be specified. Multiple label matches can be specified.

With label matching, we may or may not know the image tag we're looking
for, so update the logic to account for that, and optionally fetch image
labels, if we need to match on them.