Encapsulate manifold run.
Manifold Run doesn't work on PRs due to Travis' set up with regards to
secrets.

Manifold Run is only needed when we release the image - to inject the
docker credentials. We don't need to do this when we run tests and
linting.

Let's move it to only run when we build the images (after merging a PR)
so that we can ensure PRs still build.